feat: se implemento el uso de un archivo de barril
