+++ Cliente/Peticiones/BorrarTarea.js
@@ -1,4 +1,4 @@
-import { destroy } from "../helper/fetch.js";
+import { destroy } from "../helper/index.js";
 
 export const eliminarTareaPorId = async (nombre) => {
     // La URL será algo como: todos/15
