+++ Cliente/Peticiones/index.js
@@ -1,4 +1,4 @@
-import { request } from "../helper/fetch.js";
+import { request } from "../helper/index.js";
 
 export const BuscarNombre = async (nombre) =>{
