+++ Cliente/helper/index.js
@@ -1,0 +1,3 @@
+export * from "./put.js";
+export * from "./Delete.js";
+export * from "./fetch.js";
